requirements.txt
flask>=3.0.0
bcrypt>=4.1.2
argon2-cffi>=23.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
